Update Harlech-SimplePattern.ino
Updated with keep alive code that flips on D3 (that's connected with a resistor and transistor to ground) .
It's used to pull an extra 100mA or so for 1 second in every 10. that should help keep a USB power brick.
This ratio (1s in 10s) may need to be tweaked depending on the power brick.